// packages/ui-components/src/transaction/types.ts
export type ParamType = {
  name?: string
  type: string
  value: any
}